+++ test/cppadcg/dae_index_reduction/pantelides.cpp
@@ -53,7 +53,7 @@ TEST_F(CppADCGIndexReductionTest, PantelidesPendulum2D) {
 
     ASSERT_TRUE(reducedFun != NULL);
 
-    ASSERT_EQ(3, pantelides.getDifferentiationIndex());
+    ASSERT_EQ(size_t(3), pantelides.getDifferentiationIndex());
 
     delete fun;
     delete reducedFun;
